+++ applications/plugins/SofaCUDA/src/SofaCUDA/component/collision/response/contact/CudaPenalityContactForceField.cpp
@@ -20,7 +20,7 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 #include <sofa/gpu/cuda/CudaTypes.h>
-#include "CudaPenalityContactForceField.inl"
+#include <SofaCUDA/component/collision/response/contact/CudaPenalityContactForceField.inl>
 #include <sofa/core/ObjectFactory.h>
 
 namespace sofa::component::collision::response::contact
